requirements.txt
click
matplotlib
numpy
opencv-python
pillow